+++ OniAccess/ModConfig.cs
@@ -4,5 +4,6 @@ namespace OniAccess {
 	public class ModConfig {
 		public CoordinateMode CoordinateMode { get; set; } = CoordinateMode.Off;
 		public bool AutoMoveCursor { get; set; } = false;
+		public bool LockZoom { get; set; } = true;
 	}
 }